Update p_whiten.py
changed my chunk data type from np.float64 to np.float32.

+++ ml_ephys/preprocessing/p_whiten.py
@@ -42,7 +42,7 @@ def compute_AAt_matrix_for_chunk(num):
     t2=int(np.minimum(X.N2(),(t1+chunk_size))) # last timepoint of chunk (+1)
     
     # Ensuring that this chunk value is float64 to avoid svd complications
-    chunk=X.readChunk(i1=0,N1=X.N1(),i2=t1,N2=t2-t1).astype(np.float64) # Read the chunk
+    chunk=X.readChunk(i1=0,N1=X.N1(),i2=t1,N2=t2-t1).astype(np.float32) # Read the chunk
     
     ret=chunk @ np.transpose(chunk)
